More refactoring of file structure
