+++ to-do-list/src/Header.js
@@ -1,7 +1,8 @@
-export const Header = () => {
+export const Header = (props) => {
+    const {heading} = props;
     return (
         <div>
-            <h1>To Do List</h1>
+            <h1>{heading}</h1>
         </div>
     )
 }
